Create multiplication table.ph
Nive wants to learn multiplication tables. Write a program to print the multiplication table of a given number up to some end value given by the user. Input format The first line of the input consists of a number. The number of which you want to print its multiplication table. The second line of the input is the end value. The number up to which you want to print the table.