Create club_attendance.ph
A week day party is considered to be successful if the number of people attending this party is between 700 and 1000, both inclusive. [ The hall is neither under-crowded or overcrowded]. A weekend party is considered to be successful if the number of people attending the party is at least 1500. Can you please help Mahirl in classifying whether the given parties were successful or not? Given the day details and the number of people who have attended the party on that day, write a program to classify the party as successful or not. Input format The first line of the input consists of a string that corresponds to the day details. The 7 days are denoted as “SUN”, “MON”, “TUE”, “WED”, “THU”, “FRI” and “SAT”. Friday, Saturday and Sunday are considered to be week ends. The second line of the input consists of an integer that corresponds to the number of people who have attended the party.