Create print_non_fibonachi.ph
The find_non_fibonacci_numbers function generates the Fibonacci sequence up to the given limit using a while loop, and then checks each number up to the limit to see if it is in the Fibonacci sequence. Any number that is not in the sequence is added to a list of non-Fibonacci numbers. Finally, the function returns the list of non-Fibonacci numbers.

The example usage code prompts the user to enter the upper limit, calls the find_non_fibonacci_numbers function to generate the list of non-Fibonacci numbers, and then prints the result.