Update and rename reverse divisor.ph to reverse divisor and also check pallandrome.ph
In this code, we first take input from the user and store it in the variable s. Then we check if the string is equal to its reverse (which is obtained using the slicing operator [::-1]). If the condition is true, we print that the string is a palindrome. Otherwise, we print that it is not a palindrome.