Create cpatch_neibours_8.ph
Tim and Bob are off to a famous Education Fair "Knowledge Forum 2017" at Uzhlanda. 
This time they have to travel without their guardians. Tim got very interested in the arrangement of seats inside the train coach. 
The entire coach could be viewed as an arrangement of consecutive blocks of size 8. and the pattern is repeated for every set of 8 berths. 
Tim and Bob are playing this game of finding the co-partner in train of each berth. Write a program to do the same. Input format The input consists of an integer N,
which corresponds to the berth number whose neighbor is to be found out.
Output format The output is to display the berth of the neighbor of the corresponding seat. Refer sample input and output for formatting specifications