Create check_list_neigbour.ph
Competitive Test "Axcent Academy" has arranged for a competitive test for medical students from rural villages. Those successful students of the test will be awarded the scholarship for their NEET preparations at Axcent Academy. Benny, the co-coordinator and founder of the academy has given one problem for the first stage of the test. The problem goes like this: Given an array A1, A2, ..., AN, count the number of sub arrays of array A which are non-decreasing. A sub array A[i, j], where 1 ≤ i ≤ j ≤ N is a sequence of integers Ai, Ai+1, ..., Aj. A sub array A[i, j] is non-decreasing if Ai ≤ Ai+1 ≤ Ai+2 ≤ ... ≤ Aj. Count the total number of such sub arrays. Benny himself has not computed the solution of the problem. Write a program to help him find the answer for the same to evaluate the students. Input format The first line of input contains a single integer N denoting the size of array. Assume that the maximum value for N as 50. The second line contains N space-separated integers A1, A2, ...,AN denoting the elements of the array. Output format Output in a single line, the count of the total number of such sub arrays. Refer sample input and output for formatting specifications.