Create insert_2list_in_one.ph
Given a list of numbers in Python List, write a Python program to create a list of tuples having first element as the number and second element as the cube of the number. Input format Input contains integers separated by spaces. Output format Output displays in the form of list of tuples. Sample testcases Input 1 1 2 3 4 5 Output 1 [(1, 1), (2, 8), (3, 27), (4, 64), (5, 125)]