Create swap_arr_in_list.ph
Rajesh was going through alternative array sorting. He wishes to print the array alternatively. Hence hired you. Your task is to help rajesh in printing the array alternatively. An alternative array is an array in which first element is maximum of the whole array second element is minimum of the whole array.Third element is second largest. Fourth element is second smallest. And so on… Input format You are given with the length of array ‘n’. Next line contains ‘n’ space separated numbers.